PaX: Add PR_LOCKDOWN_MPROT_X prctl
Upon invocation, denies the current task any future executable
mmap/mprotect.

Intended to use by ld.so upon finishing mmap-ing libraries when running
as superuser, CAP_SETUID, or a blacklisted user.
The expected reason to do so is to not allow an attacker a privilege
escalation by introducing his code into a privileged task e.g. after a
buffer overflow.

Of course, invoking the prctl breaks dlopen... but this is the intended
effect.