PaX: remove libffi emutramp
The libffi is meant to provide bindings to runtime linked libraries.

With exec pages lockdown in effect, it is not possible to dlopen()
anyway (so there is nothing to FFI), and otherwise libffi should, at
least, have no problem obtaining a suitable mapping.

Remove the 'E' flag from /proc/[pid]/status, as the libffi was the only
consumer of it (the only other emutramp target being gcc nested
functions).