Remove no-op define in not-limiting-PML4-entries
The __shadow_user_pgds is itself already - correctly - ifdefed the same
way.

The ifdef place here is itself just where a crash happens with
weak-and-slow UDEREF and 47-bit user-space addresses (it then tries to
copy 512 pointers to adjanced region, not belonging to the per-cpu-PGDs
buffer).

Further corrections for commit message: The '+1' in bit shift
calculations is due to the 48-th kernel-space bit (also extended to
64-th, per Documentation/x86/x86_64/mm.txt ).
Finally, assuming L1 cache is always 4kb was probably wrong.

Fixes: 43ff48655d ("PaX: Do not limit the number of PML4 entries")